Improve image creation documentation.
Split finalise and release stages of image creation.

Build directory is now the current directory by default.

Docs generation is less noisy.

+++ makeimage.sh
@@ -2,6 +2,17 @@
 
 SECTOR_SIZE=512
 
+SRC_FILES=$(cd $(dirname $0) ; pwd -P)
+
+SRC_VERSION_LONG=$(cd $SRC_FILES && git describe)
+SRC_VERSION_SHORT=$(cd $SRC_FILES && git describe --abbrev=0)
+VERSION=${VERSION:-${SRC_VERSION_SHORT}}
+ZIP_NAME=musicbox_${VERSION}.zip
+IMG_NAME=musicbox_${VERSION}.img
+
+BUILD_DIR=${MKIMG_BUILD_DIR:-musicbox_build}
+ROOTFS_DIR=${MKIMG_ROOTFS_DIR:-${BUILD_DIR}/rootfs}
+OUTPUT_IMG=${BUILD_DIR}/${IMG_NAME}
 
 bigger() {
     local IMG_FILE=$1
@@ -48,18 +59,18 @@ smaller() {
     local NEW_SIZE_SAFE=$(expr $NEW_SIZE / $SECTOR_SIZE \* $SECTOR_SIZE)
     local IMG_SIZE=$(ls -l $IMG_FILE | cut -d" " -f5)
     if [ ! -f "$IMG_FILE" ]; then
-        echo "** ERROR: No image file found **"
-        return
+        echo "** FATAL: No image file found **"
+        exit 1
     fi
-    if [ IMG_SIZE -lt $NEW_SIZE_SAFE ]; then
-        echo "** ERROR: Requested new size ($NEW_SIZE_SAFE) is larger than current size ($IMG_SIZE) **"
-        return
+    if [ $IMG_SIZE -lt $NEW_SIZE_SAFE ]; then
+        echo "ERROR: Specifed size ($NEW_SIZE_SAFE) is larger than current size ($IMG_SIZE)"
+        return 1
     fi
-    sudo echo "Shrinking $IMG_FILE from $IMG_SIZE to $NEW_SIZE_SAFE bytes..."
+    sudo echo "INFO: Reducing $IMG_FILE from $IMG_SIZE to $NEW_SIZE_SAFE bytes..."
     #ROOT_SIZE=`used` + 100M breathing space
     local LOOP_DEV=$(sudo losetup -fP --show $IMG_FILE)
     local ROOT_PART=${LOOP_DEV}p2
-    sudo e2fsck -f ${LOOP_DEV}p2
+    sudo e2fsck -f ${ROOT_PART}
     sudo resize2fs ${ROOT_PART} $NEW_SIZE_SAFE
     cat <<EOF | sudo fdisk ${LOOP_DEV}
 d
@@ -78,38 +89,41 @@ EOF
     # TODO fix this, needs to include /boot size etc
     #truncate --size $NEW_SIZE_SAFE $IMG_FILE
 
+    IMG_SIZE=$(ls -l $IMG_FILE | cut -d" " -f5)
+    IMG_SIZE=$(expr $IMG_SIZE \/ 1024 \/ 1024)
+    echo "INFO: Reduced $IMG_FILE size is ${IMG_SIZE}MB"
     echo "** Success **"
+    return 0
 }
 
 
-release() {
+finalise() {
     local INPUT_IMG=$1
-    local SRC_FILES=$(cd $(dirname $0) ; pwd -P)
 
-    local SRC_VERSION_LONG=$(cd $SRC_FILES && git describe)
-    local SRC_VERSION_SHORT=$(cd $SRC_FILES && git describe --abbrev=0)
-    local VERSION=${VERSION:-${SRC_VERSION_SHORT}}
-    local ZIP_NAME=musicbox_${VERSION}.zip
-    local IMG_NAME=musicbox_${VERSION}.img
-
-    local BUILD_DIR=${MKIMG_BUILD_DIR:-${SRC_FILES}/_build}
-    local ROOTFS_DIR=${MKIMG_ROOTFS_DIR:-${BUILD_DIR}/rootfs}
+    if [ -f "$OUTPUT_IMG" ]; then
+        echo "INFO: Existing image found at $OUTPUT_IMG. Nothing to do."
+        return
+    fi
 
     if [ ! -f "$INPUT_IMG" ]; then
-        echo "** ERROR: No musicbox image found **"
+        echo "** FATAL: No image found at $INPUT_IMG **"
         exit 1
     fi
-    sudo echo "Info: Checking have permission to mount the disk images."
+    sudo echo "INFO: Checking permission to mount the disk images."
 
-    echo "Info: Creating $ZIP_NAME release in $BUILD_DIR from $INPUT_IMG..."
+    echo "INFO: Creating $OUTPUT_IMG from $INPUT_IMG..."
 
     rm -rf ${BUILD_DIR}
     mkdir -p ${BUILD_DIR}
-    cp $INPUT_IMG ${BUILD_DIR}/${IMG_NAME}
-    cd ${BUILD_DIR}
+    cp $INPUT_IMG $OUTPUT_IMG
 
-    local OFFSET=$(fdisk -l $IMG_NAME  | grep Linux | awk -F" "  '{ print $2 }')
-    local LOOP_DEV=$(sudo losetup -fP --show $IMG_NAME)
+    if [ ! -f "$OUTPUT_IMG" ]; then
+        echo "** FATAL: Could not create output image $OUTPUT_IMG **"
+        exit 1
+    fi
+
+    local OFFSET=$(fdisk -l $OUTPUT_IMG  | grep Linux | awk -F" "  '{ print $2 }')
+    local LOOP_DEV=$(sudo losetup -fP --show $OUTPUT_IMG)
     local ROOT_PART=${LOOP_DEV}p2
     mkdir -p ${ROOTFS_DIR}
     sudo mount ${ROOT_PART} ${ROOTFS_DIR}
@@ -132,44 +146,58 @@ release() {
     sudo umount $ROOTFS_DIR
     sudo e2fsck -fy $ROOT_PART
     sudo zerofree -v $ROOT_PART
-    sudo losetup -D $IMG_NAME
+    sudo losetup -D $OUTPUT_IMG
+    rm -rf $ROOTFS_DIR
+
+    local IMG_SIZE=$(ls -l $OUTPUT_IMG | cut -d" " -f5)
+    IMG_SIZE=$(expr $IMG_SIZE \/ 1024 \/ 1024)
+    echo "INFO: Created $OUTPUT_IMG (size: ${IMG_SIZE}MB)"
+    echo "** Success **"
+    return 0
+}
+
+
+release() {
+    finalise $1
 
     ###
     # TODO: Shrink image to fit on smaller SD cards? Who still uses 1G SD cards?!
     ###
-    #smaller $IMG_NAME $(expr 1024 \* 1024 \* 1024)
+    #smaller $OUTPUT_IMG $(expr 1024 \* 1024 \* 1024)
 
-    rm -rf $ROOTFS_DIR
+    pushd $SRC_FILES/docs
+    make text latexpdf > /dev/null
+    popd
+    cp $SRC_FILES/docs/_build/text/{changes,faq}.txt  $BUILD_DIR/
+    cp $SRC_FILES/docs/_build/latex/PiMusicBox.pdf  $BUILD_DIR/
+    pushd $BUILD_DIR
+    md5sum * > MD5SUMS
+    zip -9 $ZIP_NAME *
 
-    if [ ! -n "$IMAGE_ONLY" ]; then
-        cd $SRC_FILES/docs
-        make text latexpdf
-        cp _build/text/{changes,faq}.txt  $BUILD_DIR/
-        cp _build/latex/PiMusicBox.pdf  $BUILD_DIR/
-        cd $BUILD_DIR
-        md5sum * > MD5SUMS
-        zip -9 $ZIP_NAME *
-    fi
-
-    echo "Info: Created $ZIP_NAME release in $BUILD_DIR"
+    ZIP_SIZE=$(ls -l $ZIP_NAME | cut -d" " -f5)
+    ZIP_SIZE=$(expr $ZIP_SIZE \/ 1024 \/ 1024)
+    echo "INFO: Release $ZIP_NAME size is ${ZIP_SIZE}MB"
     echo "** Success **"
+    return 0
 }
-
 
 if [ "${BASH_SOURCE[0]}" == "${0}" ]; then
     # Not sourced
-    case "$1" in
+    case "$2" in
         bigger)
-            bigger "${@:2}"
+            bigger "$@"
             ;;
         smaller)
-            echo TODO smaller "${@:2}"
+            echo TODO smaller "$@"
+            ;;
+        finalise)
+            finalise "$@"
             ;;
         release)
-            release "${@:2}"
+            release "$@"
             ;;
         *)
-            echo "Usage: $0 bigger|smaller|release <args>"
+            echo "Usage: $0 <image> bigger|smaller|finalise|release <args>"
             ;;
     esac
 fi
